moved openai_rev code to __init__.py file
